+++ src/main/java/prismatic/shards/stellarity/registry/StellarityCriteriaTriggers.java
@@ -4,19 +4,15 @@ import net.minecraft.advancements.triggers.CriterionTrigger;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import prismatic.shards.stellarity.Stellarity;
-import prismatic.shards.stellarity.registry.advancement_criterion.SpecialCraftTrigger;
-import prismatic.shards.stellarity.registry.advancement_criterion.VoidFishedTrigger;
+import prismatic.shards.stellarity.registry.criterion_trigger.DashTrigger;
+import prismatic.shards.stellarity.registry.criterion_trigger.SpecialCraftTrigger;
+import prismatic.shards.stellarity.registry.criterion_trigger.VoidFishedTrigger;
 
 public interface StellarityCriteriaTriggers {
-	VoidFishedTrigger VOID_FISHED = register(
-		"void_fished",
-		new VoidFishedTrigger()
-	);
+	VoidFishedTrigger VOID_FISHED = register("void_fished", new VoidFishedTrigger());
+	SpecialCraftTrigger SPECIAL_CRAFT = register("special_craft", new SpecialCraftTrigger());
+	DashTrigger DASH = register("dash", new DashTrigger());
 
-	SpecialCraftTrigger SPECIAL_CRAFT = register(
-		"special_craft",
-		new SpecialCraftTrigger()
-	);
 
 	static <T extends CriterionTrigger<?>> T register(String id, T trigger) {
 		return Registry.register(BuiltInRegistries.TRIGGER_TYPES, Stellarity.id(id), trigger);